Display guichat GUI layout
The basic layout of the GUI has been completed. Further refinement
is needed, but the skeleton now exists to build in chat
functionality.

The Makefile has also been updated to reflect the slightly changed
source structure.

The .gitignore now includes a check for the "learning" folder so
that I can practice GTK and C++ concecpts without getting int the
way of the main program.